feat: use prisma as db, add in /users router with create + read operations, and more This commit is... a bit large! Here's a recap:
- Started to work on Sessions, which can work with Google, Apple, and Noelware Accounts
- Refractored authentication for Basic/Bearer (soon) tokens
- Added utilities for Argon2

+++ internal/search/engine.go
@@ -29,7 +29,7 @@ type Result struct {
 	// MaxScore returns the max score of the search itself, this can
 	// be nil if it doesn't support it. This is only in the Elasticsearch
 	// and Tsubasa engines only.
-	MaxScore *int32 `json:"max_score,omitempty"`
+	MaxScore *float64 `json:"max_score,omitempty"`
 
 	// TotalHits returns how many hits in total from the request body itself.
 	TotalHits int64 `json:"total_hits,omitempty"`
